Move website-releated to website directory, fix infinite listener bug
