+++ app.js
@@ -2,9 +2,9 @@ const Discord = require('discord.js');
 const chalk = require('chalk');
 const credentials = require('./credentials.json');
 var serverCount = 0;
-const website = require("./web.js");
+const website = require("./website/web.js");
 var io = website.io;
-var web = website.app;
+
 website.server.listen(process.env.PORT||8080);
 const manager = new Discord.ShardingManager('beemo/shard.js', {
 	totalShards: credentials.shardCount,
