+++ credentials.json
@@ -1,0 +1,1 @@
+{"token":"MjE4NTA4MDg5NDkyNjM1NjUw.C07HyA.AjRsr_JQATHMxkGUMiPgVE4GX-g","shardCount":2}
